+++ lib/countly_flutter.dart
@@ -7,6 +7,15 @@ import 'dart:io' show Platform;
 import 'package:flutter/foundation.dart';
 import 'package:pedantic/pedantic.dart';
 
+/// Attribution Keys to record indirect attribution
+/// IDFA is for iOS and AdvertisingID is for Android
+abstract class AttributionKey {
+  /// For iOS IDFA
+  static String IDFA = 'idfa';
+  /// For Android advertising ID
+  static String AdvertisingID = 'adid';
+}
+
 enum LogLevel { INFO, DEBUG, VERBOSE, WARNING, ERROR }
 
 abstract class CountlyConsent {
@@ -50,6 +59,7 @@ class Countly {
     'PRODUCTION': '0',
     'ADHOC': '2'
   };
+
   static Map<String, String> deviceIDType = {
     'TemporaryDeviceID': 'TemporaryDeviceID'
   };
@@ -1420,33 +1430,73 @@ class Countly {
   }
 
   /// Enable campaign attribution reporting to Countly.
-  /// For iOS use 'recordAttributionID' instead of 'enableAttribution'
-  /// Should be call before Countly init
+  @Deprecated('Use recordIndirectAttribution instead')
   static Future<String?> enableAttribution() async {
-    List<String> args = [];
-    final String? result = await _channel.invokeMethod(
-        'enableAttribution', <String, dynamic>{'data': json.encode(args)});
-    log(result);
-    return result;
+    log('Calling enableAttribution');
+    String error = 'enableAttribution is deprecated, use recordIndirectAttribution instead';
+    log(error);
+    return 'Error : $error';
   }
 
   /// set attribution Id for campaign attribution reporting.
-  /// Currently implemented for iOS only
-  /// For Android just call the enableAttribution to enable campaign attribution.
+  /// If this is call for iOS then 'attributionID' is IDFA
+  /// If this is call for Android then 'attributionID' is ADID
+  @Deprecated('Use recordIndirectAttribution instead')
   static Future<String?> recordAttributionID(String attributionID) async {
-    if (!Platform.isIOS) {
-      return 'recordAttributionID : To be implemented';
-    }
+    log('Calling recordAttributionID: [$attributionID]');
+    log('recordAttributionID is deprecated, use recordIndirectAttribution instead');
     if (attributionID.isEmpty) {
       String error = 'recordAttributionID, attributionID cannot be empty';
       log(error);
       return 'Error : $error';
     }
-    List<String> args = [];
-    args.add(attributionID);
+    Map<String, String> attributionValues = {};
+    if(Platform.isIOS){
+      attributionValues[AttributionKey.IDFA] = attributionID;
+    }
+    else {
+      attributionValues[AttributionKey.AdvertisingID] = attributionID;
+    }
+    final String? result = await recordIndirectAttribution(attributionValues);
+    return result;
+  }
+
+  /// set indirect attribution Id for campaign attribution reporting.
+  /// Use 'AttributionKey' to set key of IDFA and ADID
+  static Future<String?> recordIndirectAttribution(Map<String, String> attributionValues) async {
+    int attributionValuesCount = attributionValues.length;
+    log('Calling recordIndirectAttribution: [$attributionValuesCount]');
+    attributionValues.forEach((k, v) {
+      if(k.isEmpty) {
+        String error = 'recordIndirectAttribution, Key should not be empty, ignoring that key-value pair';
+        log(error);
+        attributionValues.removeWhere((key, value) => key == k && value == v);
+      }
+    });
+    List<dynamic> args = [];
+    args.add(attributionValues);
     final String? result = await _channel.invokeMethod(
-        'recordAttributionID', <String, dynamic>{'data': json.encode(args)});
-    log(result);
+        'recordIndirectAttribution', <String, dynamic>{'data': json.encode(args)});
+    return result;
+  }
+
+  /// set direct attribution Id for campaign attribution reporting.
+  /// Currently implemented for Android only.
+  static Future<String?> recordDirectAttribution(String campaignType, String campaignData) async {
+    log('Calling recordDirectAttribution: [$campaignType] with campaignData: [$campaignData]');
+    if (!Platform.isAndroid) {
+      return 'recordDirectAttribution : To be implemented';
+    }
+    if (campaignType.isEmpty) {
+      String error = 'recordDirectAttribution, campaignId cannot be empty';
+      log(error);
+      return 'Error : $error';
+    }
+    List<String> args = [];
+    args.add(campaignType);
+    args.add(campaignData);
+    final String? result = await _channel.invokeMethod(
+        'recordDirectAttribution', <String, dynamic>{'data': json.encode(args)});
     return result;
   }
 
@@ -1514,6 +1564,18 @@ class Countly {
 
       if(config.enableRemoteConfigAutomaticDownload != null) {
         countlyConfig['enableRemoteConfigAutomaticDownload'] = config.enableRemoteConfigAutomaticDownload;
+      }
+
+      if(config.daCampaignType != null) {
+        countlyConfig['campaignType'] = config.daCampaignType;
+      }
+
+      if(config.daCampaignData != null) {
+        countlyConfig['campaignData'] = config.daCampaignData;
+      }
+
+      if(config.iaAttributionValues != null) {
+        countlyConfig['attributionValues'] = config.iaAttributionValues;
       }